requirements.txt
discord.py>=2.4.0
yt-dlp>=2023.3.4
python-dotenv>=1.0.0
PyNaCl>=1.5.0  # Required for voice support 